Move categoryModal to ./Category
