+++ src/components/homepage.js
@@ -1,7 +1,7 @@
 import React, { Component } from 'react'
 import Tabs from '../components/common/common_tabs';
 
-import CategoryModal from './categoryModal'
+import CategoryModal from './Category/categoryModal'
 
 class Homepage extends Component {
